More script refactoring
