+++ qupath-gui-fx/src/main/java/qupath/lib/gui/scripting/EditableText.java
@@ -19,7 +19,7 @@
  * #L%
  */
 
-package qupath.lib.scripting.languages;
+package qupath.lib.gui.scripting;
 
 
 /**
