+++ qupath-gui-fx/src/main/java/qupath/lib/gui/scripting/syntax/GeneralCodeSyntax.java
@@ -21,8 +21,7 @@
 
 package qupath.lib.gui.scripting.syntax;
 
-import qupath.lib.scripting.languages.EditableText;
-import qupath.lib.scripting.languages.ScriptSyntax;
+import qupath.lib.gui.scripting.EditableText;
 
 /**
  * Abstract class to represent the typical syntaxes found in most programming languages.
