+++ qupath-gui-fx/src/main/java/qupath/lib/gui/scripting/syntax/GroovySyntax.java
@@ -25,7 +25,7 @@ package qupath.lib.gui.scripting.syntax;
 
 import java.util.Set;
 
-import qupath.lib.scripting.languages.EditableText;
+import qupath.lib.gui.scripting.EditableText;
 
 /**
  * Class to take care of the Groovy syntax formatting.
