+++ qupath-gui-fx/src/main/java/qupath/lib/gui/scripting/syntax/JsonSyntax.java
@@ -30,8 +30,8 @@ import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonSyntaxException;
 
+import qupath.lib.gui.scripting.EditableText;
 import qupath.lib.io.GsonTools;
-import qupath.lib.scripting.languages.EditableText;
 
 /**
  * Class that takes care of JSON syntax.
