+++ qupath-gui-fx/src/main/java/qupath/lib/gui/scripting/syntax/XmlSyntax.java
@@ -37,7 +37,7 @@ import javax.xml.transform.stream.StreamSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import qupath.lib.scripting.languages.EditableText;
+import qupath.lib.gui.scripting.EditableText;
 
 /**
  * Class that takes care of XML syntax.
